Fix build

+++ src/renderer/RenderDevice.cpp
@@ -151,8 +151,8 @@ void RenderDevice::tBGFXCallback::screenShot(
          {
             for (unsigned int i = 0; i < _height; i++)
             {
-               const uint8_t* src = static_cast<const uint8_t*>(_data) + i * _pitch;
-               uint8_t* dst = static_cast<uint8_t*>(tex->data()) + i * (_width * 4);
+               const unsigned int* src = static_cast<const unsigned int*>(_data) + i * _pitch;
+               unsigned int* dst = static_cast<unsigned int*>(tex->data()) + i * (_width * 4);
                copy_bgra_rgba<false>(dst, src, _width);
             }
          }
